Move converse-bookmark-views into folder
