+++ src/plugins/bookmark-views/templates/list.js
@@ -1,5 +1,5 @@
 import { html } from "lit-html";
-import { __ } from '../i18n';
+import { __ } from 'i18n';
 
 
 const bookmark_item = (o) => {
